requirements.txt
absl-py==2.2.1
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
astunparse==1.6.3
blinker==1.9.0
cachetools==5.5.2
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chess==1.11.2
click==8.1.8
contourpy==1.3.1
cssselect2==0.8.0
cycler==0.12.1
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
dotenv==0.9.9
executing==2.2.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.3.2
gast==0.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-genai==1.15.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
graphviz==0.20.3
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
idna==3.10
imageio==2.37.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
keras==3.9.1
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.0.8
networkx==3.4.2
numpy==2.1.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.73.0
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydot==3.0.4
pygame==2.6.1
Pygments==2.19.1
pyparsing==3.2.3
python-chess==1.999
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
rsa==4.9.1
scipy==1.15.2
seaborn==0.13.2
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
stockfish==3.28.0
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow_keras==0.1
termcolor==3.0.1
tinycss2==1.4.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
traitlets==5.14.3
triton==3.2.0
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
ultralytics==8.3.109
ultralytics-thop==2.0.14
uritemplate==4.1.1
urllib3==2.3.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
